AI Chatbot/requirements.txt
streamlit 
google-generativeai 
faiss-cpu
sentence-transformers 
pypdf2
reportlab